# src/main/java/com/example/pdffiller/controller/FileUploadController.java
package com.example.pdffiller.controller;

import java.nio.file.Path;
import java.util.concurrent.TimeUnit;
import java.util.stream.Collectors;
import java.util.stream.Stream;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.core.io.Resource;
import org.springframework.http.HttpHeaders;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder;
import org.springframework.web.servlet.mvc.support.RedirectAttributes;

import com.example.pdffiller.storage.StorageFileNotFoundException;
import com.example.pdffiller.storage.StorageService;

@Controller
public class FileUploadController {

	private final StorageService storageService;

	@Autowired
	public FileUploadController(StorageService storageService) {
		this.storageService = storageService;
	}

	@GetMapping("/")
	public String listUploadedFiles(Model model) {

/*		try(Stream<Path> allFiles = storageService.loadAll()) {
			model.addAttribute("files", allFiles.map(
							path -> MvcUriComponentsBuilder.fromMethodName(FileUploadController.class,
									"serveFile", path.getFileName().toString()).build().toUri().toString())
					.collect(Collectors.toList()));
		}*/

		try(Stream<Path> allFiles = storageService.loadAll()) {
			model.addAttribute("files", allFiles.map(
							path -> MvcUriComponentsBuilder.fromMethodName(PDFController.class,
									"getPdfFields", path.getFileName().toString()).build().toUri().toString())
					.collect(Collectors.toList()));
		}




		return "uploadForm";
	}

	@GetMapping("/files/{filename:.+}")
	@ResponseBody
	public ResponseEntity<Resource> serveFile(@PathVariable String filename) {

		Resource file = storageService.loadAsResource(filename);
		return ResponseEntity.ok().header(HttpHeaders.CONTENT_DISPOSITION,
				"attachment; filename=\"" + file.getFilename() + "\"").body(file);
	}

	@PostMapping("/")
	public String handleFileUpload(@RequestParam("file") MultipartFile file,
								   RedirectAttributes redirectAttributes) {

		storageService.store(file);

		// Some delay because of Not found error on beginning
		try {
			TimeUnit.MILLISECONDS.sleep(1000);
		} catch (InterruptedException ie) {
			Thread.currentThread().interrupt();
		}

		redirectAttributes.addFlashAttribute("message",
				"You successfully uploaded " + file.getOriginalFilename() + "!");
		return "redirect:/";
	}

	@ExceptionHandler(StorageFileNotFoundException.class)
	public ResponseEntity<?> handleStorageFileNotFound(StorageFileNotFoundException exc) {
		return ResponseEntity.notFound().build();
	}

}